app/requirements.txt
Click==7.0
Flask==1.1.2
Jinja2==2.10.1
MarkupSafe==1.1.0
mongoengine==0.16.2
pymongo==3.7.2
six~=1.15.0
uWSGI==2.0.17.1
Werkzeug==0.15.0
pymongo==3.7.2
Flask-PyMongo==2.2.0
flasgger==0.9.0
google-cloud-automl==0.10.0
google-cloud-language==2.0.0
google-cloud-automl==0.10.0
google-cloud-vision
google-cloud-storage
pandas==1.2.4
mongoengine-goodjson==1.1.6
Flask-JWT==0.3.2
tensorflow==2.5.0

